15:fix minor issues with db and pagination

+++ commons/src/main/java/org/gs/incode/services/stringtransformation/dtos/PagedResponse.java
@@ -6,11 +6,11 @@ import lombok.Value;
 @Value
 public class PagedResponse<T> {
   private final List<T> content;
-  private final int totalElements;
+  private final long totalElements;
   private final int page;
   private final int size;
 
-  public PagedResponse(List<T> content, int totalElements, int page, int size) {
+  public PagedResponse(List<T> content, long totalElements, int page, int size) {
     this.content = content;
     this.totalElements = totalElements;
     this.page = page;
